# backend/docs/swagger.yaml
basePath: /
definitions:
...
  handler.APIResponse:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
...
  model.CustomerLevel:
    properties:
      createdAt:
        type: string
      id:
        type: integer
      name:
        type: string
      sort:
        type: integer
    type: object
  model.CustomerLevelRequest:
    properties:
      name:
        type: string
      sort:
        type: integer
    required:
    - name
    type: object
...
  model.CustomerSource:
    properties:
      createdAt:
        type: string
      id:
        type: integer
      name:
        type: string
      sort:
        type: integer
    type: object
  model.CustomerSourceRequest:
    properties:
      name:
        type: string
      sort:
        type: integer
    required:
    - name
    type: object
...
  model.SystemSettingsResponse:
    properties:
      contractNumberPrefix:
        type: string
      customerLevels:
        items:
          $ref: '#/definitions/model.CustomerLevel'
        type: array
      customerLimit:
        type: integer
      customerSources:
        items:
          $ref: '#/definitions/model.CustomerSource'
        type: array
      dealDropDays:
        type: integer
      followUpDropDays:
        type: integer
      holidayModeEnabled:
        type: boolean
      showFullContact:
        type: boolean
    type: object
  model.UpdateSystemSettingsRequest:
    properties:
      contractNumberPrefix:
        type: string
      customerLimit:
        type: integer
      dealDropDays:
        type: integer
      followUpDropDays:
        type: integer
      holidayModeEnabled:
        type: boolean
      showFullContact:
        type: boolean
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: 后端服务API文档
  title: Backend API
  version: "1.0"
paths:
...
  /api/v1/settings:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.SystemSettingsResponse'
      security:
      - BearerAuth: []
      summary: 获取系统设置
      tags:
      - 系统设置
    put:
      consumes:
      - application/json
      parameters:
      - description: 系统设置
        in: body
        name: settings
        required: true
        schema:
          $ref: '#/definitions/model.UpdateSystemSettingsRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: 更新系统设置
      tags:
      - 系统设置
  /api/v1/settings/customer-levels:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.CustomerLevel'
            type: array
      security:
      - BearerAuth: []
      summary: 获取客户级别列表
      tags:
      - 系统设置
    post:
      consumes:
      - application/json
      parameters:
      - description: 客户级别
        in: body
        name: level
        required: true
        schema:
          $ref: '#/definitions/model.CustomerLevelRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.CustomerLevel'
      security:
      - BearerAuth: []
      summary: 创建客户级别
      tags:
      - 系统设置
  /api/v1/settings/customer-levels/{id}:
    delete:
      parameters:
      - description: 级别ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: 删除客户级别
      tags:
      - 系统设置
    put:
      consumes:
      - application/json
      parameters:
      - description: 级别ID
        in: path
        name: id
        required: true
        type: integer
      - description: 客户级别
        in: body
        name: level
        required: true
        schema:
          $ref: '#/definitions/model.CustomerLevelRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: 更新客户级别
      tags:
      - 系统设置
  /api/v1/settings/customer-sources:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.CustomerSource'
            type: array
      security:
      - BearerAuth: []
      summary: 获取客户来源列表
      tags:
      - 系统设置
    post:
      consumes:
      - application/json
      parameters:
      - description: 客户来源
        in: body
        name: source
        required: true
        schema:
          $ref: '#/definitions/model.CustomerSourceRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.CustomerSource'
      security:
      - BearerAuth: []
      summary: 创建客户来源
      tags:
      - 系统设置
  /api/v1/settings/customer-sources/{id}:
    delete:
      parameters:
      - description: 来源ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: 删除客户来源
      tags:
      - 系统设置
    put:
      consumes:
      - application/json
      parameters:
      - description: 来源ID
        in: path
        name: id
        required: true
        type: integer
      - description: 客户来源
        in: body
        name: source
        required: true
        schema:
          $ref: '#/definitions/model.CustomerSourceRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: 更新客户来源
      tags:
      - 系统设置
...
  /api/v1/users/avatar/upload:
    post:
      consumes:
      - multipart/form-data
      description: 上传用户头像到阿里云OSS
      parameters:
      - description: 头像文件（支持 JPG、PNG、WEBP，最大 2MB）
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/handler.APIResponse'
            - properties:
                data:
                  additionalProperties:
                    type: string
                  type: object
              type: object
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/handler.APIResponse'
        "401":
          description: 未登录或登录已失效
          schema:
            $ref: '#/definitions/handler.APIResponse'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/handler.APIResponse'
      security:
      - BearerAuth: []
      summary: 上传用户头像
      tags:
      - users
...
securityDefinitions:
  BearerAuth:
    description: 'Bearer token, 格式: "Bearer {token}"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"